Add mouse-over events
Add mouse enter and mouse leave events for the earth and moon sprites.
Note - it doesn't seem to trigger new events if the mouse remains still,
for example if the mouse enters the moon then remains still, a "mouse
leave" event won't be fired when the moon moves out from under the
mouse. Only when the mouse moves will the game "notice" that it has left
the moon